+++ frontend/src/data/featuredVideos.ts
@@ -1,0 +1,24 @@
+export const FEATURED_VIDEOS = [
+  'BV1hWi7B4EQ8',
+  'BV1ncBNBrE8z',
+  'BV11o2UBhErL',
+  'BV1LMvDBXEMB',
+  'BV11kiMBpE4T',
+  'BV1APBaBsEua',
+  'BV1b7BZB6E7w',
+  'BV16BB5BTE4u',
+  'BV164BMBnETA',
+  'BV1fQBTB5EK8',
+  'BV1vaBLBcEq2',
+  'BV1a9vtBXEuy',
+  'BV1jviTBuErm',
+  'BV11dBMBhE1J',
+  'BV1c3qDBBEa2',
+  'BV1r4SWBdEjg',
+  'BV1bxYfzjE9Q',
+  'BV15zBQBwEcb',
+  'BV15m1vBTET3',
+  'BV19FvdBYEXN',
+]
+
+export const buildBilibiliUrl = (bvid: string) => `https://www.bilibili.com/video/${bvid}`